+++ Time/time_now_ms.cpp
@@ -23,11 +23,11 @@ void    time_reset_clock_now_hook(void)
     return ;
 }
 
-long    time_now_ms(void)
+int64_t    time_now_ms(void)
 {
     std::chrono::system_clock::time_point time_now;
     std::chrono::milliseconds milliseconds;
-    long long milliseconds_count;
+    int64_t milliseconds_count;
 
     if (g_time_now_ms_hook != ft_nullptr)
         time_now = g_time_now_ms_hook();
@@ -35,16 +35,16 @@ long    time_now_ms(void)
         time_now = std::chrono::system_clock::now();
     milliseconds = std::chrono::duration_cast<std::chrono::milliseconds>(time_now.time_since_epoch());
     milliseconds_count = milliseconds.count();
-    if (milliseconds_count > static_cast<long long>(LONG_MAX))
+    if (milliseconds_count > LONG_MAX)
     {
         (void)(FT_ERR_OUT_OF_RANGE);
         return (LONG_MAX);
     }
-    if (milliseconds_count < static_cast<long long>(LONG_MIN))
+    if (milliseconds_count < LONG_MIN)
     {
         (void)(FT_ERR_OUT_OF_RANGE);
         return (LONG_MIN);
     }
     (void)(FT_ERR_SUCCESS);
-    return (static_cast<long>(milliseconds_count));
+    return (milliseconds_count);
 }
